Revert "feat(core): execute plugins in isolated processes (#21760)"
This reverts commit c2581dcf0b5a605a9a74e7bf6911bc8e93210170.
