+++ packages/devkit/src/utils/convert-nx-executor.spec.ts
@@ -1,7 +1,4 @@
-import { requireNx } from '../../nx';
 import { convertNxExecutor } from './convert-nx-executor';
-
-const { workspaceRoot } = requireNx();
 
 describe('Convert Nx Executor', () => {
   it('should convertNxExecutor to builder correctly and produce the same output', async () => {
@@ -9,14 +6,12 @@ describe('Convert Nx Executor', () => {
     const { schema } = require('@angular-devkit/core');
     const {
       TestingArchitectHost,
-      // nx-ignore-next-line
-    } = require('@angular-devkit/architect/testing') as typeof import('@angular-devkit/architect/testing');
+    } = require('@angular-devkit/architect/testing');
     const { Architect } = require('@angular-devkit/architect');
 
     const registry = new schema.CoreSchemaRegistry();
     registry.addPostTransform(schema.transforms.addUndefinedDefaults);
     const testArchitectHost = new TestingArchitectHost();
-    testArchitectHost.workspaceRoot = workspaceRoot;
     const architect = new Architect(testArchitectHost, registry);
 
     const convertedExecutor = convertNxExecutor(echoExecutor);
