+++ packages/devkit/src/utils/convert-nx-executor.ts
@@ -8,7 +8,6 @@ const {
   Workspaces,
   readNxJsonFromDisk,
   retrieveProjectConfigurationsWithAngularProjects,
-  shutdownPluginWorkers,
 } = requireNx();
 
 /**
@@ -39,7 +38,6 @@ export function convertNxExecutor(executor: Executor) {
             (workspaces as any).readProjectsConfigurations({
               _includeProjectsFromAngularJson: true,
             });
-      shutdownPluginWorkers?.();
 
       const context: ExecutorContext = {
         root: builderContext.workspaceRoot,
