+++ packages/nx/src/adapter/angular-json.ts
@@ -2,7 +2,7 @@ import { existsSync } from 'fs';
 import * as path from 'path';
 import { readJsonFile } from '../utils/fileutils';
 import { ProjectsConfigurations } from '../config/workspace-json-project-json';
-import { NxPluginV2 } from '../project-graph/plugins';
+import { NxPluginV2 } from '../utils/nx-plugin';
 
 export const NX_ANGULAR_JSON_PLUGIN_NAME = 'nx-angular-json-plugin';
 
@@ -15,8 +15,6 @@ export const NxAngularJsonPlugin: NxPluginV2 = {
     }),
   ],
 };
-
-export default NxAngularJsonPlugin;
 
 export function shouldMergeAngularProjects(
   root: string,
