+++ packages/nx/src/config/workspaces.spec.ts
@@ -3,7 +3,6 @@ import { TempFs } from '../internal-testing-utils/temp-fs';
 import { withEnvironmentVariables } from '../internal-testing-utils/with-environment';
 import { retrieveProjectConfigurations } from '../project-graph/utils/retrieve-workspace-files';
 import { readNxJson } from './configuration';
-import { shutdownPluginWorkers } from '../project-graph/plugins/plugin-pool';
 
 const libConfig = (root, name?: string) => ({
   name: name ?? toProjectName(`${root}/some-file`),
@@ -49,11 +48,10 @@ describe('Workspaces', () => {
 
       const { projects } = await withEnvironmentVariables(
         {
-          NX_WORKSPACE_ROOT_PATH: fs.tempDir,
+          NX_WORKSPACE_ROOT: fs.tempDir,
         },
         () => retrieveProjectConfigurations(fs.tempDir, readNxJson(fs.tempDir))
       );
-      await shutdownPluginWorkers();
       expect(projects['my-package']).toEqual({
         name: 'my-package',
         root: 'packages/my-package',
@@ -67,7 +65,6 @@ describe('Workspaces', () => {
           },
         },
       });
-      await shutdownPluginWorkers();
     });
   });
 });
