+++ packages/nx/src/devkit-exports.ts
@@ -47,19 +47,16 @@ export { workspaceLayout } from './config/configuration';
 
 export type {
   NxPlugin,
+  NxPluginV1,
   NxPluginV2,
+  ProjectTargetConfigurator,
   CreateNodes,
   CreateNodesFunction,
   CreateNodesResult,
   CreateNodesContext,
   CreateDependencies,
   CreateDependenciesContext,
-} from './project-graph/plugins';
-
-export type {
-  NxPluginV1,
-  ProjectTargetConfigurator,
-} from './utils/nx-plugin.deprecated';
+} from './utils/nx-plugin';
 
 /**
  * @category Workspace
