+++ packages/nx/src/plugins/js/lock-file/pnpm-parser.spec.ts
@@ -11,7 +11,7 @@ import {
   ProjectGraphBuilder,
   RawProjectGraphDependency,
 } from '../../../project-graph/project-graph-builder';
-import { CreateDependenciesContext } from '../../../project-graph/plugins';
+import { CreateDependenciesContext } from '../../../utils/nx-plugin';
 
 jest.mock('fs', () => {
   const memFs = require('memfs').fs;
