+++ packages/nx/src/plugins/js/project-graph/build-dependencies/explicit-project-dependencies.ts
@@ -6,7 +6,7 @@ import {
 import { join, relative } from 'path';
 import { workspaceRoot } from '../../../../utils/workspace-root';
 import { normalizePath } from '../../../../utils/path';
-import { CreateDependenciesContext } from '../../../../project-graph/plugins';
+import { CreateDependenciesContext } from '../../../../utils/nx-plugin';
 import {
   RawProjectGraphDependency,
   validateDependency,
