+++ packages/nx/src/plugins/package-json-workspaces/create-nodes.ts
@@ -8,42 +8,48 @@ import { toProjectName } from '../../config/workspaces';
 import { readJsonFile, readYamlFile } from '../../utils/fileutils';
 import { combineGlobPatterns } from '../../utils/globs';
 import { NX_PREFIX } from '../../utils/logger';
+import { NxPluginV2 } from '../../utils/nx-plugin';
 import { output } from '../../utils/output';
 import {
   PackageJson,
   readTargetsFromPackageJson,
 } from '../../utils/package-json';
 import { joinPathFragments } from '../../utils/path';
-import { workspaceRoot } from '../../utils/workspace-root';
-import { CreateNodes } from '../../project-graph/plugins';
 
-const readJson = (f) => readJsonFile(join(workspaceRoot, f));
-const patterns = getGlobPatternsFromPackageManagerWorkspaces(
-  workspaceRoot,
-  readJson
-);
-const negativePatterns = patterns.filter((p) => p.startsWith('!'));
-const positivePatterns = patterns.filter((p) => !p.startsWith('!'));
-if (
-  // There are some negative patterns
-  negativePatterns.length > 0 &&
-  // No positive patterns
-  (positivePatterns.length === 0 ||
-    // Or only a single positive pattern that is the default coming from root package
-    (positivePatterns.length === 1 && positivePatterns[0] === 'package.json'))
-) {
-  positivePatterns.push('**/package.json');
+export function getNxPackageJsonWorkspacesPlugin(root: string): NxPluginV2 {
+  const readJson = (f) => readJsonFile(join(root, f));
+  const patterns = getGlobPatternsFromPackageManagerWorkspaces(root, readJson);
+
+  // If the user only specified a negative pattern, we should find all package.json
+  // files and only return those that don't match a negative pattern.
+  const negativePatterns = patterns.filter((p) => p.startsWith('!'));
+  let positivePatterns = patterns.filter((p) => !p.startsWith('!'));
+
+  if (
+    // There are some negative patterns
+    negativePatterns.length > 0 &&
+    // No positive patterns
+    (positivePatterns.length === 0 ||
+      // Or only a single positive pattern that is the default coming from root package
+      (positivePatterns.length === 1 && positivePatterns[0] === 'package.json'))
+  ) {
+    positivePatterns.push('**/package.json');
+  }
+
+  return {
+    name: 'nx/core/package-json-workspaces',
+    createNodes: [
+      combineGlobPatterns(positivePatterns),
+      (p) => {
+        if (!negativePatterns.some((negative) => minimatch(p, negative))) {
+          return createNodeFromPackageJson(p, root);
+        }
+        // A negative pattern matched, so we should not create a node for this package.json
+        return {};
+      },
+    ],
+  };
 }
-export const createNodes: CreateNodes = [
-  combineGlobPatterns(positivePatterns),
-  (p, _, { workspaceRoot }) => {
-    if (!negativePatterns.some((negative) => minimatch(p, negative))) {
-      return createNodeFromPackageJson(p, workspaceRoot);
-    }
-    // A negative pattern matched, so we should not create a node for this package.json
-    return {};
-  },
-];
 
 export function createNodeFromPackageJson(pkgJsonPath: string, root: string) {
   const json: PackageJson = readJsonFile(join(root, pkgJsonPath));
