+++ packages/nx/src/plugins/package-json-workspaces/index.ts
@@ -1,2 +1,1 @@
 export * from './create-nodes';
-export const name = 'nx/core/package-json-workspaces';
