+++ packages/nx/src/plugins/project-json/build-nodes/package-json-next-to-project-json.spec.ts
@@ -3,7 +3,7 @@ import * as memfs from 'memfs';
 import '../../../internal-testing-utils/mock-fs';
 
 import { PackageJsonProjectsNextToProjectJsonPlugin } from './package-json-next-to-project-json';
-import { CreateNodesContext } from '../../../project-graph/plugins';
+import { CreateNodesContext } from '../../../utils/nx-plugin';
 const { createNodes } = PackageJsonProjectsNextToProjectJsonPlugin;
 
 describe('nx project.json plugin', () => {
