+++ packages/nx/src/plugins/target-defaults/target-defaults-plugin.ts
@@ -8,7 +8,7 @@ import {
 } from '../../config/workspace-json-project-json';
 import { readJsonFile } from '../../utils/fileutils';
 import { combineGlobPatterns } from '../../utils/globs';
-import { NxPluginV2 } from '../../project-graph/plugins';
+import { NxPluginV2 } from '../../utils/nx-plugin';
 import {
   PackageJson,
   readTargetsFromPackageJson,
@@ -126,8 +126,6 @@ export const TargetDefaultsPlugin: NxPluginV2 = {
     },
   ],
 };
-
-export default TargetDefaultsPlugin;
 
 function getExecutorToTargetMap(
   packageJsonTargets: Record<string, TargetConfiguration>,
