+++ packages/nx/src/project-graph/affected/locators/project-glob-changes.spec.ts
@@ -1,7 +1,7 @@
 import { ProjectGraphProjectNode } from '../../../config/project-graph';
 import { ProjectConfiguration } from '../../../config/workspace-json-project-json';
 
-import * as nxPlugin from '../../../project-graph/plugins';
+import * as nxPlugin from '../../../utils/nx-plugin';
 import { DeletedFileChange } from '../../file-utils';
 import { getTouchedProjectsFromProjectGlobChanges } from './project-glob-changes';
