+++ packages/nx/src/project-graph/affected/locators/project-glob-changes.ts
@@ -1,16 +1,23 @@
 import { TouchedProjectLocator } from '../affected-project-graph-models';
 import { minimatch } from 'minimatch';
 import { workspaceRoot } from '../../../utils/workspace-root';
+import { getNxRequirePaths } from '../../../utils/installation-directory';
 import { join } from 'path';
 import { existsSync } from 'fs';
 import { configurationGlobs } from '../../utils/retrieve-workspace-files';
-import { loadNxPlugins } from '../../plugins/internal-api';
+import { loadNxPlugins } from '../../../utils/nx-plugin';
 import { combineGlobPatterns } from '../../../utils/globs';
 
 export const getTouchedProjectsFromProjectGlobChanges: TouchedProjectLocator =
   async (touchedFiles, projectGraphNodes, nxJson): Promise<string[]> => {
     const globPattern = combineGlobPatterns(
-      configurationGlobs(await loadNxPlugins(nxJson?.plugins, workspaceRoot))
+      configurationGlobs(
+        await loadNxPlugins(
+          nxJson?.plugins,
+          getNxRequirePaths(workspaceRoot),
+          workspaceRoot
+        )
+      )
     );
 
     const touchedProjects = new Set<string>();
