--- packages/nx/src/project-graph/plugins/index.ts
@@ -1,6 +1,0 @@
-export * from './public-api';
-
-export {
-  readPluginPackageJson,
-  registerPluginTSTranspiler,
-} from './worker-api';
