+++ packages/nx/src/project-graph/project-graph.ts
@@ -17,6 +17,7 @@ import {
   retrieveWorkspaceFiles,
 } from './utils/retrieve-workspace-files';
 import { readNxJson } from '../config/nx-json';
+import { unregisterPluginTSTranspiler } from '../utils/nx-plugin';
 
 /**
  * Synchronously reads the latest cached copy of the workspace's ProjectGraph.
@@ -103,6 +104,8 @@ export async function buildProjectGraphAndSourceMapsWithoutDaemon() {
     )
   ).projectGraph;
   performance.mark('build-project-graph-using-project-file-map:end');
+
+  unregisterPluginTSTranspiler();
 
   return { projectGraph, sourceMaps };
 }
