+++ packages/nx/src/utils/logger.ts
@@ -31,11 +31,6 @@ export const logger = {
   fatal: (...s) => {
     console.error(...s);
   },
-  verbose: (...s) => {
-    if (process.env.NX_VERBOSE_LOGGING) {
-      console.log(...s);
-    }
-  },
 };
 
 export function stripIndent(str: string): string {
