+++ packages/nx/src/utils/plugins/plugin-capabilities.ts
@@ -1,19 +1,19 @@
+import { workspaceRoot } from '../workspace-root';
 import * as chalk from 'chalk';
 import { dirname, join } from 'path';
-
-import { ProjectConfiguration } from '../../config/workspace-json-project-json';
-import { readPluginPackageJson } from '../../project-graph/plugins';
-import { RemotePlugin } from '../../project-graph/plugins/internal-api';
-import { loadRemoteNxPlugin } from '../../project-graph/plugins/plugin-pool';
-import { readJsonFile } from '../fileutils';
-import { getNxRequirePaths } from '../installation-directory';
 import { output } from '../output';
-import { PackageJson } from '../package-json';
-import { getPackageManagerCommand } from '../package-manager';
-import { workspaceRoot } from '../workspace-root';
-import { hasElements } from './shared';
-
 import type { PluginCapabilities } from './models';
+import { hasElements } from './shared';
+import { readJsonFile } from '../fileutils';
+import { getPackageManagerCommand } from '../package-manager';
+import {
+  loadNxPluginAsync,
+  NxPlugin,
+  readPluginPackageJson,
+} from '../nx-plugin';
+import { getNxRequirePaths } from '../installation-directory';
+import { PackageJson } from '../package-json';
+import { ProjectConfiguration } from '../../config/workspace-json-project-json';
 
 function tryGetCollection<T extends object>(
   packageJsonPath: string,
@@ -101,17 +101,24 @@ async function tryGetModule(
   packageJson: PackageJson,
   workspaceRoot: string,
   projects: Record<string, ProjectConfiguration>
-): Promise<RemotePlugin | null> {
+): Promise<NxPlugin | null> {
   try {
     return packageJson.generators ??
       packageJson.executors ??
       packageJson['nx-migrations'] ??
       packageJson['schematics'] ??
       packageJson['builders']
-      ? await loadRemoteNxPlugin(packageJson.name, workspaceRoot)
+      ? (
+          await loadNxPluginAsync(
+            packageJson.name,
+            getNxRequirePaths(workspaceRoot),
+            projects,
+            workspaceRoot
+          )
+        ).plugin
       : ({
           name: packageJson.name,
-        } as RemotePlugin);
+        } as NxPlugin);
   } catch {
     return null;
   }
